+++ src/inspect_ai/approval/__init__.py
@@ -3,7 +3,7 @@ from ._approval import Approval, ApprovalDecision
 from ._approver import Approver
 from ._auto import auto_approver
 from ._human.approver import human_approver
-from ._policy import ApprovalPolicy
+from ._policy import ApprovalPolicy, policy_approver
 from ._registry import approver
 
 __all__ = [
@@ -13,6 +13,7 @@ __all__ = [
     "ApprovalPolicy",
     "approval",
     "approver",
+    "policy_approver",
     "human_approver",
     "auto_approver",
 ]